Fix UT Entropy_BitstreamWriter.write_symbol_no_update, value br.allow_update_cdf = 0; not set
