+++ Source/Lib/Decoder/Codec/EbDecParseFrame.h
@@ -163,8 +163,18 @@ void parse_super_block(EbDecHandle *dec_handle, ParseCtxt *parse_ctxt, uint32_t 
 void svt_tile_init(TileInfo *cur_tile_info, FrameHeader *frame_header, int32_t tile_row,
                    int32_t tile_col);
 
-EbErrorType init_svt_reader(SvtReader *r, const uint8_t *data, const uint8_t *data_end,
-                            const size_t read_size, uint8_t allow_update_cdf);
+static int read_is_valid(const uint8_t *start, size_t len, const uint8_t *end) {
+    return len != 0 && len <= (size_t)(end - start);
+}
+
+static INLINE EbErrorType init_svt_reader(SvtReader *r, const uint8_t *data, const uint8_t *data_end,
+                            const size_t read_size, uint8_t allow_update_cdf) {
+    if (read_is_valid(data, read_size, data_end) && !svt_reader_init(r, data, read_size))
+        r->allow_update_cdf = allow_update_cdf;
+    else
+        return EB_Corrupt_Frame;
+    return EB_ErrorNone;
+}
 
 EbErrorType start_parse_tile(EbDecHandle *dec_handle_ptr, ParseCtxt *parse_ctxt,
                              TilesInfo *tiles_info, int tile_num, int is_mt);
